+++ src/main/java/com/github/aytchell/qrgen/ColorConfig.java
@@ -8,22 +8,26 @@ import java.awt.image.BufferedImage;
 public class ColorConfig {
     ArgbValue onColor;
     ArgbValue offColor;
-    ArgbValue markerColor;
+    ArgbValue outerMarkerColor;
+    ArgbValue innerMarkerColor;
 
     public ColorConfig(ArgbValue onColor, ArgbValue offColor) {
-        this.onColor = onColor;
-        this.offColor = offColor;
-        this.markerColor = onColor;
+        this(onColor, offColor, onColor);
     }
 
     public ColorConfig(ArgbValue onColor, ArgbValue offColor, ArgbValue markerColor) {
+        this(onColor, offColor, markerColor, markerColor);
+    }
+
+    public ColorConfig(ArgbValue onColor, ArgbValue offColor, ArgbValue outerMarkerColor, ArgbValue innerMarkerColor) {
         this.onColor = onColor;
         this.offColor = offColor;
-        this.markerColor = markerColor;
+        this.outerMarkerColor = outerMarkerColor;
+        this.innerMarkerColor = innerMarkerColor;
     }
 
     public int determineImageType() {
-        if (onColor.hasAlpha() || offColor.hasAlpha() || markerColor.hasAlpha()) {
+        if (onColor.hasAlpha() || offColor.hasAlpha() || outerMarkerColor.hasAlpha() || innerMarkerColor.hasAlpha()) {
             return BufferedImage.TYPE_INT_ARGB;
         }
         return BufferedImage.TYPE_INT_RGB;